Upgradability mechanisms script version (#1313)
* ''

* ''

* ''

* Changes genesis block version to 0.

* a

* a

* All tests are done.

* All tests passed for changed block version from int32 to uint16

* Adds validation of rejecting blocks with unknown versions.

* Changes txn version from int32 to uint16.

* .

* Adds comments to exported functions.

* Change functions name from ConvertFromRpcScriptPubKeyToRPCScriptPubKey to ConvertFromAppMsgRPCScriptPubKeyToRPCScriptPubKey and from ConvertFromRPCScriptPubKeyToRpcScriptPubKey to ConvertFromRPCScriptPubKeyToAppMsgRPCScriptPubKey

* change comment to "ScriptPublicKey represents a Kaspad ScriptPublicKey"

* delete part (tx.Version < 0) that cannot be exist on the if statement.

* Revert protobuf version.

* Fix a comment.

* Fix a comment.

* Rename a variable.

* Rename a variable.

* Remove a const.

* Rename a type.

* Rename a field.

* Rename a field.

* Remove commented-out code.

* Remove dangerous nil case in DomainTransactionOutput.Clone().

* Remove a constant.

* Fix a string.

* Fix wrong totalScriptPubKeySize in transactionMassStandalonePart.

* Remove a constant.

* Remove an unused error.

* Fix a serialization error.

* Specify version types to be uint16 explicitly.

* Use constants.ScriptPublicKeyVersion.

* Fix a bad test.

* Remove some whitespace.

* Add a case to utxoEntry.Equal().

* Rename scriptPubKey to scriptPublicKey.

* Remove a TODO.

* Rename constants.

* Rename a variable.

* Add version to parseShortForm.

+++ domain/consensus/utils/blockheader/blockheader.go
@@ -3,7 +3,7 @@ package blockheader
 import "github.com/kaspanet/kaspad/domain/consensus/model/externalapi"
 
 type blockHeader struct {
-	version              int32
+	version              uint16
 	parentHashes         []*externalapi.DomainHash
 	hashMerkleRoot       *externalapi.DomainHash
 	acceptedIDMerkleRoot *externalapi.DomainHash
@@ -25,7 +25,7 @@ func (bh *blockHeader) SetTimeInMilliseconds(timeInMilliseconds int64) {
 	bh.timeInMilliseconds = timeInMilliseconds
 }
 
-func (bh *blockHeader) Version() int32 {
+func (bh *blockHeader) Version() uint16 {
 	return bh.version
 }
 
@@ -124,7 +124,7 @@ func (bh *blockHeader) ToMutable() externalapi.MutableBlockHeader {
 
 // NewImmutableBlockHeader returns a new immutable header
 func NewImmutableBlockHeader(
-	version int32,
+	version uint16,
 	parentHashes []*externalapi.DomainHash,
 	hashMerkleRoot *externalapi.DomainHash,
 	acceptedIDMerkleRoot *externalapi.DomainHash,
